# blog-admin/src/main/java/com/zhang/blogadmin/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "shenkunlin@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/wpCommentmeta/{id}": {
      "delete": {
        "tags": [
          "WpCommentmetaController"
        ],
        "summary": "根据ID删除WpCommentmeta",
        "description": "根据ID删除WpCommentmeta方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除WpCommentmeta方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除WpCommentmeta"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "WpCommentmetaController"
        ],
        "summary": "根据ID修改WpCommentmeta",
        "description": "根据ID修改WpCommentmeta方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改WpCommentmeta方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "WpCommentmeta",
            "description": "传入WpCommentmeta的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpCommentmeta"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改WpCommentmeta"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "WpCommentmetaController"
        ],
        "summary": "根据ID查询WpCommentmeta",
        "description": "根据ID查询WpCommentmeta方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改WpCommentmeta方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改WpCommentmeta"
            ,
            "schema": {
              "$ref": "#/definitions/Result«WpCommentmeta»"
            }
          }
        }
      }
    },
    "/wpCommentmeta/search": {
      "post": {
        "tags": [
          "WpCommentmetaController"
        ],
        "summary": "不带分页条件搜索WpCommentmeta",
        "description": "不带分页条件搜索WpCommentmeta方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "WpCommentmeta",
            "description": "传入WpCommentmeta的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpCommentmeta"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索WpCommentmeta"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpCommentmeta»»"
            }
          }
        }
      }
    },
    "/wpCommentmeta/search/{page}/{size}": {
      "get": {
        "tags": [
          "WpCommentmetaController"
        ],
        "summary": "分页列表查询WpCommentmeta",
        "description": "分页列表查询WpCommentmeta方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索WpCommentmeta"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpCommentmeta»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "WpCommentmetaController"
        ],
        "summary": "分页条件搜索WpCommentmeta",
        "description": "分页条件搜索WpCommentmeta方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "WpCommentmeta",
            "description": "传入WpCommentmeta的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpCommentmeta"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索WpCommentmeta"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpCommentmeta»»"
            }
          }
        }
      }
    },
...
    "/wpComments/{id}": {
      "delete": {
        "tags": [
          "WpCommentsController"
        ],
        "summary": "根据ID删除WpComments",
        "description": "根据ID删除WpComments方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除WpComments方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除WpComments"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "WpCommentsController"
        ],
        "summary": "根据ID修改WpComments",
        "description": "根据ID修改WpComments方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改WpComments方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "WpComments",
            "description": "传入WpComments的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpComments"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改WpComments"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "WpCommentsController"
        ],
        "summary": "根据ID查询WpComments",
        "description": "根据ID查询WpComments方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改WpComments方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改WpComments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«WpComments»"
            }
          }
        }
      }
    },
    "/wpComments/search": {
      "post": {
        "tags": [
          "WpCommentsController"
        ],
        "summary": "不带分页条件搜索WpComments",
        "description": "不带分页条件搜索WpComments方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "WpComments",
            "description": "传入WpComments的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpComments"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索WpComments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpComments»»"
            }
          }
        }
      }
    },
    "/wpComments/search/{page}/{size}": {
      "get": {
        "tags": [
          "WpCommentsController"
        ],
        "summary": "分页列表查询WpComments",
        "description": "分页列表查询WpComments方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索WpComments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpComments»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "WpCommentsController"
        ],
        "summary": "分页条件搜索WpComments",
        "description": "分页条件搜索WpComments方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "WpComments",
            "description": "传入WpComments的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpComments"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索WpComments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpComments»»"
            }
          }
        }
      }
    },
...
    "/wpLinks/{id}": {
      "delete": {
        "tags": [
          "WpLinksController"
        ],
        "summary": "根据ID删除WpLinks",
        "description": "根据ID删除WpLinks方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除WpLinks方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除WpLinks"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "WpLinksController"
        ],
        "summary": "根据ID修改WpLinks",
        "description": "根据ID修改WpLinks方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改WpLinks方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "WpLinks",
            "description": "传入WpLinks的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpLinks"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改WpLinks"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "WpLinksController"
        ],
        "summary": "根据ID查询WpLinks",
        "description": "根据ID查询WpLinks方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改WpLinks方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改WpLinks"
            ,
            "schema": {
              "$ref": "#/definitions/Result«WpLinks»"
            }
          }
        }
      }
    },
    "/wpLinks/search": {
      "post": {
        "tags": [
          "WpLinksController"
        ],
        "summary": "不带分页条件搜索WpLinks",
        "description": "不带分页条件搜索WpLinks方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "WpLinks",
            "description": "传入WpLinks的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpLinks"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索WpLinks"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpLinks»»"
            }
          }
        }
      }
    },
    "/wpLinks/search/{page}/{size}": {
      "get": {
        "tags": [
          "WpLinksController"
        ],
        "summary": "分页列表查询WpLinks",
        "description": "分页列表查询WpLinks方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索WpLinks"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpLinks»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "WpLinksController"
        ],
        "summary": "分页条件搜索WpLinks",
        "description": "分页条件搜索WpLinks方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "WpLinks",
            "description": "传入WpLinks的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/WpLinks"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索WpLinks"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpLinks»»"
            }
          }
        }
      }
    },
    "/wpOptions": {
      "get": {
        "tags": [
          "WpOptionsController"
        ],
        "summary": "查询所有WpOptions",
        "description": "查询所有WpOptions方法详情",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "查询所有WpOptions"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«WpOptions»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "WpOptionsController"
        ],
        "summary": "添加WpOptions",
        "description": "添加WpOptions方法详情",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "添加WpOptions方法详情",
            "required": true
              ,
              "schema": {
                "$ref": "#/definitions/WpOptions"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "添加WpOptions"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "Result«WpCommentmeta»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/WpCommentmeta"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "WpCommentmeta": {
      "type": "object",
      "properties": {
        "metaId": {
          "type": "long",
          "description": ""
        },
        "commentId": {
          "type": "long",
          "description": ""
        },
        "metaKey": {
          "type": "string",
          "description": ""
        },
        "metaValue": {
          "type": "string",
          "description": ""
        }
      },
      "description": "WpCommentmeta"
    },
...
    "WpOptions": {
      "type": "object",
      "properties": {
        "optionId": {
          "type": "long",
          "description": ""
        },
        "optionName": {
          "type": "string",
          "description": ""
        },
        "optionValue": {
          "type": "string",
          "description": ""
        },
        "autoload": {
          "type": "string",
          "description": ""
        }
      },
      "description": "WpOptions"
    },
...
  },
...
}